renamed utils file to more specific storybookUtils
