+++ packages/web-app/src/modules/storefront-views/blocks/StorefrontBlock.stories.tsx
@@ -1,6 +1,6 @@
 import { storiesOf } from '@storybook/react'
 import { StorefrontBlockComponent } from '../../storefront/models'
-import { getMockStrapiBlockData } from '../utils'
+import { getMockStrapiBlockData } from '../storybookUtils'
 import { StorefrontContentBlock, StorefrontHeroBlock, StorefrontRewardBlock } from './'
 
 storiesOf('Modules/Storefront/Blocks', module)
